# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Inkwell API
  version: "0.2"
  description: |
    API-first Markdown publishing platform.

    All write endpoints require `X-Api-Key`. Reads are public unless stated
    otherwise. See [docs/API.md](./API.md) for the full narrative reference.
  contact:
    name: Inkwell
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-railway-domain.up.railway.app
    description: Production (replace with actual domain)
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Shared admin key (`INKWELL_API_KEY`) or a scoped token
        (`ink_<prefix>_<secret>`).
  schemas:
...
    BacklinksResponse:
      type: object
      required: [backlinks, mentions]
      properties:
        backlinks:
          type: array
          items:
            type: object
            required: [slug, title]
            properties:
              slug:
                type: string
              title:
                type: string
              snippet:
                type: string
                nullable: true
        mentions:
          type: array
          items:
            type: object
            required: [sourceUrl]
            properties:
              sourceUrl:
                type: string
                format: uri
    GraphNode:
      type: object
      required: [slug, title]
      properties:
        slug:
          type: string
        title:
          type: string
    GraphEdge:
      type: object
      required: [sourceSlug, targetSlug]
      properties:
        sourceSlug:
          type: string
        targetSlug:
          type: string
    GraphResponse:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
              example: "Missing or invalid API key."
            requestId:
              type: string
              format: uuid
              description: Correlation ID. Present on every in-request error.
            slug:
              type: string
              description: Present on 409 Conflict for duplicate slug.
            allow:
              type: string
              description: Present on 405 Method Not Allowed.
              example: "GET, POST"
  responses:
...
    RejectedReadCredential:
      description: >
        An `X-Api-Key` was presented and rejected (revoked token, unknown prefix,
        hash mismatch, a value that is neither the shared key nor token-shaped, or
        a duplicated/non-ASCII header). Treat this as "re-authenticate", not
        "empty result" — see ADR 0015. A rejected `inkwell_session` cookie is NOT
        reported here: that channel downgrades to anonymous and returns `200` with
        published content, so a stale browser cookie never breaks public pages. A
        valid credential lacking the `read` scope also stays `200`, published only.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Missing or invalid API key."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
    ServiceUnavailable:
      description: >
        The credential could not be judged because the token lookup failed against
        Postgres (ADR 0015). Distinct from `401` — the credential was never
        actually rejected — and from `500`. Retry; check `GET /readyz`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "Service temporarily unavailable. Retry shortly."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
...
    NotFound:
      description: Resource does not exist (or caller cannot see it).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              message: "No document with slug \"hello-world\"."
              requestId: "a3b5c7d9-1234-4abc-8def-0123456789ab"
...
paths:
...
  /documents/{slug}/backlinks:
    get:
      summary: Backlinks and Webmentions for a document
      operationId: documentBacklinks
      tags: [Documents]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Internal backlinks and verified inbound Webmentions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacklinksResponse"
              example:
                backlinks:
                  - slug: "other-note"
                    title: "Other Note"
                    snippet: "…see also [[hello-world]] for context…"
                mentions:
                  - sourceUrl: "https://external.example/post"
        "404":
          $ref: "#/components/responses/NotFound"
  /documents/{slug}/graph:
    get:
      summary: One-hop neighborhood graph for a document
      operationId: documentGraph
      tags: [Documents]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: Nodes and directed edges within one hop of this note.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /graph:
    get:
      summary: Whole-garden link graph
      operationId: gardenGraph
      tags: [Documents]
      responses:
        "401":
          $ref: "#/components/responses/RejectedReadCredential"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "200":
          description: All nodes and directed edges in the bounded garden graph.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphResponse"
...